chore: organize queries files
