+++ core/queries/task/delete-task.query.ts
